requirements.txt
hydra-core
dvc
pytest
ruamel.yaml==0.17.17
great_expectations
pandas
scikit-learn
matplotlib
seaborn
zenml[server]
apache-airflow==2.7.3
feast
pendulum
shyaml
mlflow
giskard
evidently
Flask
flask-session2
psycopg2-binary
--extra-index-url https://download.pytorch.org/whl/cu121
torch
gradio